+++ tests/test_stationary_reloader_perk.py
@@ -4,6 +4,7 @@ from crimson.gameplay import (
     GameplayState,
     player_update,
 )
+from crimson.math_parity import f32
 from crimson.perks import PerkId
 from crimson.sim.input import PlayerInput
 from crimson.sim.state_types import PlayerState
@@ -28,5 +29,5 @@ def test_stationary_reloader_triples_reload_speed() -> None:
     player_update(base_player, PlayerInput(), dt=0.1, state=state)
     player_update(perk_player, PlayerInput(), dt=0.1, state=state)
 
-    assert_float_close(base_player.reload_timer, 0.9)
-    assert_float_close(perk_player.reload_timer, 0.7)
+    assert_float_close(base_player.reload_timer, f32(0.9))
+    assert_float_close(perk_player.reload_timer, f32(0.7))
